Number Puzzle
This is a 2d terminal-based Number Puzzle game with an unlimited number of columns and rows. You select the number of rows, the number of columns, and the shuffle which actually means difficulty.  The more shuffle, the more difficult it will be. But, without a higher number of rows and columns, shuffles don't worth it.
Btw, select the number of rows and columns based on your monitor size. If your monitor size is small, like mehh, RIP 🥲